Finished some driver work
The ADA102 LED driver is now finished and working

The DS3231 is also working correctly.  It isn't complete as interacting with the alarms isn't started

TelnetServer was tested and is now working as well.  Likely room for improvement, but currently functional.

IniFile is a new thing to allow reading a config file from SD.