Building up the string class
Adding a class to handle binary outputs.  The main goal for it will be to act as driver for servos and other devices controlled via PWM.

Began adding code documentation to some files.

String Class gained a lot of functionality.  All of which was tested.  Still needs Split and Regex support.