PWM output driver
Discovered that the servos in use required a specific PWM frequency and duty cycle range to work.  Gettign this in the arduino was inconsistent, so added a PWM expansion driver to cover this.

Completed the base functionality in the binary output class.
Added some new arduino device constants in Common Utils.