Driver Complete RasPi I2C no good
Finished the driver for the MCP23017 GPIO expander.  Tested it using Arduino so its working.
The RasPi I2C implementation does not appear to work.  It compiles, but doesn't appear to communicate with the peripheral.  Need to get a logic analyzer on that to see if the bus is engaged or not.