Button Display project
Tryign to tie hardware together in a project using a 4x20 display and a button pad.  These both operate off of GPIO.

Necessitated an implementation of the time interface in Raspberry Pi.

Also created a driver specific to the 3v OLED 4x20 character display in use.

Working on driver functions for the keypad at present.