Created Basic driver
Setup a project to build and test the driver for this PWM output controller.

Found a few compile problems in the I2C and UART general interface files.  The Return enum was sometimes named Returns.  This was fixed in the general interface libs as well as raspberry pi implementations.

Probably still issues in the other implementations.