Timing issues
Had to abaondone the Ras Pi project since it could not do the required timing.

Switched to arduino where there was no OS to get in the way.  Added a 100 nanoSecond delay function but something isn't right with the output on the pin.  Need a logic analyzer to sort that out.