Adding Libraries
Added in the general interface libraries as well as the Nucleo implementations.
Needed to updated the Nucleo implementations of UART and I2C to use correct variable names.
Learned that the GPIO names applied in the cube throw off all generic arrays made in the header file.  Working on a new scheme to work around that issue, or at least get this working with the current applied names.
The names are incorrect, all are GPI none have the GPO tag so the directions aren't documented in code.