Interface work
The GPIO interface was missing a port capabilities value, so that was added.  This was carried into the Nucleo implementations
Also added a Nucleo implementation of the Time interface.  Has a define to specify HAL or FreeRTOS as the time source.
Updated the library folders with the modified files.

Next up the SPI interface