Spi Caps test and watchdog
Added SPI capabilities expectations to the APA102 driver.  This allowed the board support header to test for those at compile time.
Also added some basic wathdog functions into the time interface.  Included the available support in the Nucleo implementation and added that to the application code instead of the HAL call.