Terminal commands
Improved the handling of the built in terminal commands.  Verified that they are working as well.
Created a task to more frequently poll the serial port.  Its unclear how fast characters will be coming in.  Since the ST provides no buffer for the loaded characters its possible some will be missed.  It may be necessary to explore the interrupt method of receiving the input.