Updating WiznetIface project
Bringing this project up to the latest libraries.  Starting some investigation into the incorrect behavior seen in the wiznet device.

It looks like the data on the SPI bus is being interpretted incorrectly.  The device is sending random responses to commands that should have constant output.  If the SPI lines are floating in some way maybe its getting bad commands?  Need to investigate pull resistors on that bus.