Closed socket issues
When a socket closes abruptly, there doesn't seem to be a good way to detect it on receive.  Sending might give better errors?  Application level ping/response mechanisms are most reliable for this.

Not great for the telnet connection since we'd need the OS level keep alive for it, which is defaulted to every 2 hours.

Otherwise data is getting to and from the Terminal client through the sockets.  There is some issue in processing the data that I need to check into.  Possibly fixed in the UART version that isn't ported here.