I2C Slave
Finished the General Interface template of the I2C slave.
Implemented it in STM32 HAL.  Had to modify the Cube MX file and regenerate the project to verify all the configurations.
Everything tested find but some error handling options may be missing.  Need to work more with it as is to be sure.
Also more documentation is needed.  Specifically what settings are necessary within the Cube MX to make this work.