Driver tested
Tested out the driver and it seems to work fine.  The sensor sends bad data about 40% of the time.  its near zero so it seems like it can be filtered out.

The arduino program was used to collect and view the data from the sensor.
Next up is to finish off and order the PCB